add .env functionality, error handling, renaming files based on convention, moving structs to dto folder, removing transaction objs used in repo and postgres folders
