+++ .env
@@ -1,0 +1,9 @@
+PORT=8080
+
+
+# DB Variables
+DB_HOST="localhost"
+DB_PORT=5432
+DB_USER="postgres"
+DB_PASSWORD="1234"
+DB_NAME="rozgarlink"
