+++ go.mod
@@ -3,6 +3,8 @@ module github.com/harsh-jagtap-josh/RozgarLink
 go 1.23.4
 
 require (
+	github.com/caarlos0/env/v11 v11.3.1
 	github.com/gorilla/mux v1.8.1
+	github.com/joho/godotenv v1.5.1
 	github.com/lib/pq v1.10.9
 )
