+++ internal/api/router.go
@@ -1,6 +1,8 @@
 package api
 
 import (
+	"net/http"
+
 	"github.com/gorilla/mux"
 	"github.com/harsh-jagtap-josh/RozgarLink/internal/app"
 )
@@ -8,8 +10,8 @@ import (
 func NewRouter(deps app.Dependencies) *mux.Router {
 	router := mux.NewRouter()
 
-	WorkersRouter := router.PathPrefix("/workers").Subrouter()
-	WorkersRouter.HandleFunc("/{worker_id}", HandleWorkerByID(deps.WorkerService)).Methods("GET")
+	workersRouter := router.PathPrefix("/workers").Subrouter()
+	workersRouter.HandleFunc("/{worker_id}", HandleWorkerByID(deps.WorkerService)).Methods(http.MethodGet)
 
 	return router
 }
