+++ internal/app/worker/service.go
@@ -3,6 +3,7 @@ package worker
 import (
 	"context"
 
+	"github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/dto"
 	"github.com/harsh-jagtap-josh/RozgarLink/internal/repository"
 )
 
@@ -11,7 +12,7 @@ type service struct {
 }
 
 type Service interface {
-	GetWorkerByID(ctx context.Context, tx repository.Transaction, productID int64) (repository.Worker, error)
+	GetWorkerByID(ctx context.Context, productID int64) (dto.Worker, error)
 }
 
 func NewService(workerRepo repository.WorkerStorer) Service {
@@ -20,14 +21,14 @@ func NewService(workerRepo repository.WorkerStorer) Service {
 	}
 }
 
-func (ws *service) GetWorkerByID(ctx context.Context, tx repository.Transaction, productID int64) (repository.Worker, error) {
-	workerInfoDB, err := ws.workerRepo.GetWorkerByID(ctx, tx, productID)
+func (ws *service) GetWorkerByID(ctx context.Context, productID int64) (dto.Worker, error) {
+	workerInfoDB, err := ws.workerRepo.GetWorkerByID(ctx, productID)
 	if err != nil {
-		return repository.Worker{}, nil
+		return dto.Worker{}, err
 	}
 
 	if workerInfoDB.ID == 0 {
-		return repository.Worker{}, nil
+		return dto.Worker{}, nil
 	}
 
 	return workerInfoDB, nil
