+++ internal/pkg/dto/sector.go
@@ -1,0 +1,7 @@
+package dto
+
+type Sector struct {
+	ID          int    `json:"id"`
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}
