+++ internal/repository/base.go
@@ -1,4 +1,4 @@
-package postgres
+package repository
 
 import "database/sql"
