+++ internal/repository/postgres/worker.go
@@ -4,24 +4,25 @@ import (
 	"context"
 	"database/sql"
 
+	"github.com/harsh-jagtap-josh/RozgarLink/internal/pkg/dto"
 	"github.com/harsh-jagtap-josh/RozgarLink/internal/repository"
 )
 
 type workerStore struct {
-	BaseRepository
+	repository.BaseRepository
 }
 
 func NewWorkerRepo(db *sql.DB) repository.WorkerStorer {
 	return &workerStore{
-		BaseRepository: BaseRepository{db},
+		BaseRepository: repository.BaseRepository{db},
 	}
 }
 
-func (ws *workerStore) GetWorkerByID(ctx context.Context, tx repository.Transaction, workerID int64) (repository.Worker, error) {
+func (ws *workerStore) GetWorkerByID(ctx context.Context, workerID int64) (dto.Worker, error) {
 	query := `SELECT * FROM Workers WHERE id = $1;`
 	row := ws.DB.QueryRow(query, workerID)
 
-	var worker repository.Worker
+	var worker dto.Worker
 	err := row.Scan(&worker.ID, &worker.Name, &worker.ContactNumber, &worker.Email, &worker.Gender,
 		&worker.Password, &worker.Sectors, &worker.Skills, &worker.Location, &worker.IsAvailable,
 		&worker.Rating, &worker.TotalJobsWorked, &worker.CreatedAt, &worker.UpdatedAt, &worker.Language)
